@@ -61,7 +61,7 @@ dependencies:
   reactive_forms: ^17.0.1
   json_serializable: ^6.8.0
   mask_text_input_formatter: ^2.9.0
-  icons_launcher: ^2.1.7
+  icons_launcher: ^3.0.0
   image_picker: ^1.1.2
 
   # -- media -- #
@@ -74,11 +74,11 @@ dependencies:
   # -- localization -- #
   flutter_localizations:
     sdk: flutter
-  slang: ^3.31.2
-  slang_flutter: ^3.31.0
+  slang: ^4.2.1
+  slang_flutter: ^4.2.0
 
   # -- ui -- #
-  flex_color_scheme: ^7.3.1
+  flex_color_scheme: ^8.0.1
   tab_container: ^3.5.3
   dotted_border: ^2.1.0
   auto_size_text: ^3.0.0
@@ -101,7 +101,7 @@ dev_dependencies:
   # -- testing -- #
   flutter_test:
     sdk: flutter
-  flutter_lints: ^4.0.0
+  flutter_lints: ^5.0.0
 
   # -- utils -- #
   flutter_gen_runner: ^5.3.2
@@ -110,7 +110,7 @@ dev_dependencies:
   # -- codegen -- #
   build_runner: ^2.4.11
   mobx_codegen: ^2.6.0+1
-  slang_build_runner: ^3.31.0
+  slang_build_runner: ^4.2.0
 
 flutter:
   uses-material-design: true
