@@ -3,13 +3,21 @@ import 'package:flutter/material.dart';
 import 'package:mama/src/data.dart';
 
 class LearnMoreWidget extends StatelessWidget {
-  const LearnMoreWidget({super.key});
+  final VoidCallback onPressClose;
+  final VoidCallback onPressButton;
+  final String title;
+
+  const LearnMoreWidget({
+    super.key,
+    required this.onPressClose,
+    required this.onPressButton,
+    required this.title,
+  });
 
   @override
   Widget build(BuildContext context) {
     final ThemeData themeData = Theme.of(context);
     final TextTheme textTheme = themeData.textTheme;
-
     return DecoratedBox(
       decoration: BoxDecoration(
           borderRadius: BorderRadius.circular(16), color: AppColors.lightBlue),
@@ -20,7 +28,7 @@ class LearnMoreWidget extends StatelessWidget {
             Align(
               alignment: Alignment.topRight,
               child: IconButton(
-                onPressed: () {},
+                onPressed: () => onPressClose(),
                 icon: const Icon(
                   Icons.close,
                   color: AppColors.greyColor,
@@ -28,7 +36,7 @@ class LearnMoreWidget extends StatelessWidget {
               ),
             ),
             AutoSizeText(
-              '${t.feeding.learnMoreText}:',
+              '$title:',
               style: textTheme.labelLarge?.copyWith(
                   fontWeight: FontWeight.w400,
                   color: AppColors.greyBrighterColor),
@@ -37,9 +45,9 @@ class LearnMoreWidget extends StatelessWidget {
             CustomButton(
               isSmall: false,
               type: CustomButtonType.outline,
-              onTap: () {},
+              onTap: () => onPressButton(),
               icon: IconModel(iconPath: Assets.icons.icLearnMore),
-              title: t.feeding.learnMoreBtn,
+              title: t.trackers.learnMoreBtn,
               contentPadding:
                   const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
               textStyle: textTheme.titleLarge?.copyWith(fontSize: 10),
